+++ public/locales/en/account.json
@@ -6,9 +6,11 @@
     "export": "Export {{dataType}}",
     "funding-chart": "Funding Chart",
     "health-contributions": "Health Contributions",
+    "init-health-contribution": "Init Health Contribution",
     "init-health-contributions": "Init Health Contributions",
     "liabilities": "Liabilities",
     "lifetime-volume": "Lifetime Trade Volume",
+    "maint-health-contribution": "Maint Health Contribution",
     "maint-health-contributions": "Maint Health Contributions",
     "no-data": "No data to display",
     "no-pnl-history": "No PnL History",
